Move cluster-api/machinecontroller/cloud to cluster-api/cloud
