+++ cluster-api/cloud/google/gceproviderconfig/types.go
