+++ cluster-api/cloud/google/gceproviderconfig/v1alpha1/register.go
@@ -4,7 +4,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
-	"k8s.io/kube-deploy/cluster-api/machinecontroller/cloud/google/gceproviderconfig"
+	"k8s.io/kube-deploy/cluster-api/cloud/google/gceproviderconfig"
 )
 
 const GroupName = "gceproviderconfig"
