+++ cluster-api/cloud/google/templates.go
